ptxed: do not reset the decoder on psb
When decoding PSB, we reset the decoder state to mimick the situation
right after synchronizing the decoder.  In combination with reading
ahead in order to signal events, this causes the decoder to discard
valid state.

Do not reset the decoder state when decoding PSB.  We still reset it
when decoding OVF and after synchronizing.

Change-Id: Ia806e932c420e120e7a1895f385dc10b0c5ec88c
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>