libipt: model enable state on query start
When starting the query interface after sync'ing, the decoder's flags are
clear indicating that tracing is enabled.

This is not correct; tracing is only enabled if we got a FUP in PSB+;
fix it.  Since we require a full IP after sync, it suffices to check
whether we currently have an IP.

Change-Id: Ib0a46b7bf34f248dabad4eb11240fa48c243df72
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>