all: enable libipt to be used from c++
Rename field pt_packet_unknown.private into pt_packet_unkown.priv,
because this conflicts with the C++ private keyword.

Trailing commas in enums are also not accepted by all compilers in the
most strict mode, so remove all trailing commas in all enums all over
the place.

This enables usage of libipt from a C++ program, but it is not enough to
build libipt with C++.

Fixes #2.

Change-Id: I3f15fe5af149de758302b018a00f4898610abc0f
Signed-off-by: Christian Himpel <christian.himpel@intel.com>