pt* tools: don't auto-detect cpu by default
Until now, the tools either auto-detect via cpuid the current cpu, or
let them overwrite with a family/model/stepping on the command-line.

This commit adds support for two special words for each of the tools,
"none" and "auto". "none" is the default and means the pt stream should
be encoded and decoded according to the spec, "auto" can be specified to
let the tool auto-detect via cpuid the current cpu (old-behavior).

Change-Id: I4819bd640b9f98b26d3f083a843b947ed8ec24ec