libipt: change interface for decoding unknown packets
Pass a pt_packet_unknown instead of a pt_packet that should be filled by
the decode function. And instead of passing a pt_decoder, pass its
pt_config, the current position in the buffer and a context.

Also fill default values in the extract_unknown function and extend the
tests.

Change-Id: Id35586c345e1671716d0a3e69ad9f6825ed0e9d8
Signed-off-by: Christian Himpel <christian.himpel@intel.com>