libipt: improve insn information in case of errors
Ptxed uses the ip field of the supposed-to-be next instruction for error
reporting.  At the moment, this is the IP of the last successfully decoded
instruction.

The IP is available already before we try to read the instruction's memory and
decode it.  For errors in those areas, we could already report the correct IP.

Fill in the instruction fields as early as possible to improve error reporting.

Change-Id: I1b3e2946e581a7912b037d9254b5f165fadb02c3
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>