libipt: bdm64
BDM64: An Incorrect LBR or Intel(R) Processor Trace Packet May Be
       Recorded Following a Transactional Abort.

       Use of Intel(R) Transactional Synchronization Extensions (Intel(R) TSX)
       may result in a transactional abort.  If an abort occurs immediately
       following a branch instruction, an incorrect branch target may be
       logged in an LBR (Last Branch Record) or in an Intel(R) Processor Trace
       (Intel(R) PT) packet before the LBR or Intel PT packet produced by the
       abort.

If we get a TSX abort event at a branch instruction and we can't reach
the TSX event IP from there without further PT support, we assume that
the branch destination was bogus and that the TSX abort happened directly
after executing the branch.

This does not handle the case where we can still reach the TSX event IP from
a bogus branch destination since we can't distinguish this case from a
perfectly normal TSX abort.

Change-Id: I009ed8220b05ea2b8b4e982aa91d63a567586b6e
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>