libipt: be prepared
Add a struct pt_errata and a respective field to pt_config.
Add an API function for filling in a pt_errata object based on a pt_cpu and
call that function in all tools.

The pt_errata struct is empty and the API function does nothing, but we want
to be prepared.

Change-Id: I21e3a18b92c49a05aacbb330f9eccf06d0e9b5aa
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>