pttc: update labels
The current functionality allows labels for addresses to be inserted and the
address to be looked up based on the label's name.

If we want to add (artificial) labels for things like section sizes, it will
be convenient if we were able to update the respective label as we compute
the size.

Add a function l_find() that returns a pointer to a label given its name.

Change-Id: Id9522ee74f92c0ebe672b21eb84f20ba182cecbb
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>