libipt, image: support multiple address spaces
When tracing more than one process, the traced image needs to be changed when
the process context is switched.  Add support for specifying multiple images
for different processes.

Add a struct pt_asid to identify an address-space when specifying the traced
image with calls to pt_insn_add_file() et.al.

Maintain the current pt_asid in the pt_insn_decoder based on the PIP packet and
use it when reading memory.

Change-Id: I92cac1bcbd5e301ac95084afa6bcb2ff839f558a
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>