libipt, ptxed: export pt_image
Allow the traced image to be changed while decoding.

This allows the image to be populated in advance, it allows multiple decoders
to use the same image (provided it doesn't change during decode), and it allows
switching between images.

This allows ptxed to accept --pt, --raw, and --elf in any order.

Change-Id: I1246685b41fcf529d21fe35bee5b404a0dc9a5fb
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>