libipt: change error when tracing is still disabled after event processing
pt_insn_next() starty by processing all events that do not need a decoded
instruction.  This should include everything between a trace disabled and
the subsequent trace enabled event.  If we still end up with tracing disabled
after this initial event processing phase, something is wrong.

We used to return -pte_internal in this case, which is not entirely correct.
Change this to return -pte_bad_context, which is a better fit.

Change-Id: I396c4a36a39fc4745bb82ed7f21a8493f4d8878b
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>