ptunit-selftest: avoid stack address escapes
The tests pfun_init_fail and frun_fini_fail use a stack-allocated test
fixture, provoke a ptunit error, and check that the returned error pointer
matches the test fixture's address.  In case this test fails, the address
of the stack-allocated test fixture is returned in the error report.

It's only used for error reporting ... let's avoid it nevertheless.

Change-Id: I7b61fdaf964e81ba05c7c0d57b2e83bab3c0fa83
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>