libipt: prepare for future events
Users of libipt must agree with libipt on the size of struct pt_event.  Should
we ever want to extend it, we'll be in trouble.

Add a new parameter @size to pt_qry_event() that the caller must set to
sizeof(struct pt_event) and provide at most that many bytes.

New events that need more space will not be known to old users so it shouldn't
matter if they are truncated.

Old libipt versions will only provide that portion of struct pt_event that is
actually used by the events they may report, thus leaving part of an extended
struct pt_event uninitialized.  This should be fine, as well.

Change-Id: I368e50dd84c3ecaab2687c248751853e34375f04
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>