libipt: prepare for insn extensions
Users of libipt must agree with libipt on the size of struct pt_insn.  Should
we ever want to extend it, we'll be in trouble.

Add a new parameter @size to pt_insn_next() that the caller must set to
sizeof(struct pt_insn) and provide at most that many bytes.

A future libipt that extended struct pt_insn will only provide the initial
part that is understood by its user.

An older libipt will zero out any trailing bytes it can't provide.  This may
still leave a few bit-fields cleared that the future user expects but that the
old libipt can't provide.  Users would need to check the library version to
determine if those bits are valid.

Change-Id: I0507e5f0343b18ccb08a49ccdd08dc0c5e9578ad
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>