libipt: avoid bit-field operation for instruction decode
During instruction decode, we assign the decoder's current speculative state
to the decoded instruction.  This involves bit-field operations for reading
and writing.

Given that the instruction is zero-initialized, we only need to assign this
field when it is set.  This should further be the exception, not the rule.

Change it into an if statement.

Change-Id: I72b162b69e07b7513180ae1225fefb0d704fcfc3
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>