libipt: remove pt_configure()
We provide a function to automatically determine a suitable configuration.
Unfortunately, we can't access most of the relevant information.  All we
can really do is detect the CPU and fill in the errata.

This further only works for the case where we decode on the same processor
on which we collected the trace.  Most users of libipt therefore need to
implement CPU detection for the remote case, anyway.

Replace the pt_configure() function with a static inline pt_config_init()
function to zero-initialize a configuration and fill in the size field.

Change-Id: Ica2a969b86f0556df010ce5e58c9bba44c9cda5e
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>