libipt: postpone read-ahead insn decoder errors
After decoding the current instruction, the instruction flow decoder will
proceed to the next instruction and process some events that are associated
with the next instruction's IP.  This is necessary for indicating some events
like tracing disabled already in the current instruction.

Any errors are indicated to the user immediately.  This will implicitly drop
the current instruction, which we could decode without error.  Postpone such
errors until the next iteration.

Change-Id: I695b6a1ec440086f74f993d9450f8ddde20ab245
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>