libipt: time-stamp events
Add a new field to struct pt_event to hold the time stamp count of the event.
Add a new field to struct pt_event to indicate whether timing information is
available and hence the time field is valid.

This, unfortunately, breaks backwards compatibility since we extend the
non-variant part of an event.  To make further breakages less likely, already
reserve a few more bytes for future extensions.

Change-Id: I71393ae6797312fd889ede62198ba9cc604ef011
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>