libipt: reserve errata bits
Reserve a few bytes in struct pt_errata.  The errata field is currently at
the end of struct pt_config.  As long as it remains there, we're fine and
we wouldn't really need to reserve errata bits.  Once we add new config
fileds, though, we will need them.  Let's reserve them now so we don't forget.

Change-Id: Ifdee4345babcdba65a49d447eedbd5238592ec9f
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>