libipt: change insn flow decoder's IP change logic
The instruction flow decoder does not allow it's internal IP to be changed
after the current instruction has been decoded.

The current logic is to allow the IP to be changed during the first round of
event processing.  This prevents event processing outside of pt_insn_next().

Change it to explicitly disallow IP changes during 'after' and 'peek' event
processing.

Change-Id: I6387f311cb0f5540d3fc35e2a06683cf703058c9
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>