libipt: prepare for packet extensions
Users of libipt must agree with libipt on the size of struct pt_packet.  Should
we ever want to extend it, we'll be in trouble.

Add a new parameter @size to pt_pkt_next() that the caller must set to
sizeof(struct pt_packet) and provide at most that many bytes.

A future libipt that extended struct pt_packet will only provide the initial
part that is understood by its user.

An older libipt will not know about new, bigger packets and thus won't provide
them.

Change-Id: I15ddccfdbfcc79e70a7ed730dde1c37f344713dd
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>