libipt: allow image section overlaps
When adding a new file section that would overlap one or more existing sections,
pt_image_add_file() returns an error.  It is left to the user to first remove
the existing section(s), then add the new section, plus any remaining parts of
the removed sections.

Change this to have pt_image_add_file() automatically split, shrink, or remove
existing sections that would overlap with a newly added section.

This change is not completely backwards compatible since we now allow adding
overlapping sections instead of failing with an error.  On the other hand, the
behaviour produces the expected result of adding the new file without impacting
anything that lies outside.

Change-Id: I5add3b6b07ad9e9238b03558368654d286e9eb21
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>