libipt: retry PSB+ decode
When failing to decode the PSB+ header, the decoder position will be somewhere
inside the header.

In order to resume decoding after an error, we need to either store the decode
context (i.e. header or non-header) or reset the decoder state to the PSB packet
and attempt to decode the full PSB+ header again.  Let's do the latter.

Change-Id: I9ef3dac333e87e9111835678b18047446bde05a3
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>