libipt: report read-ahead errors
All queries read ahead in order to report upcoming events.  Any errors during
reading ahead are suppressed assuming that we will run into the same error
again on the next query.

This allows us to decode as far as possible before we have to give up due to
the decode error.  On the other hand, such a decode error may hide an upcoming
event.  In this case, we might actually decode too far.

Report read-ahed errors immediately to avoid this.  In case of errors, decode
will stop one event earlier than strictly necessary.

One may argue whether reaching the end of the trace is an error in this sense.
It may as well hide an upcoming event.  On the other hand, without an explicit
trace disabled event at the end of the trace, we don't know how far to decode,
anyway.  And we already indicate the upcoming end of the trace in advance.
Let's exclude it.

Change-Id: I7fc6b113c25e346ad499025c6aa177b718df9176
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>