libipt, ild: fix a0-a3 MOV decode
The width of the displacement operand of the a0-a3 MOV instruction depends on
the address-size, not the operand-size.  Fix the respective case in
compute_disp_dec() and add helper functions to determine the effective address
size.

We seem to have gotten this wrong also in our ild unit tests.  Fix them and add
some more.

Change-Id: I1dc593834b270bca3fa54c754b69f994e86ce673
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>