libipt: ignore zero-displacement calls for return compression
A near call with zero displacement, i.e. to the next instruction, is used to
get the current code address for position independent code.  Such calls are
to be ignored when constructing the return-address stack for return compression.

This causes frequent decode errors when decoding the trace for shared libraries
on 32-bit systems.  Fix it.

Change-Id: I4a5360888c2d8049c5d0e950735e94e5520aad06
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>