libipt, ild: combine imm_bytes checks
We're comparing the length for immediate operands one-by-one.  As long as we can
guarantee that imm2_bytes is always zero if imm1_bytes is zero, we can combine
all checks into one.

Since imm2_bytes is zero-initialized and only set if also imm1_bytes is set,
we're fine.  Combine the checks.

Change-Id: I9ed28be0e19bdac3f96f4a5480750deb798cd22e
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>