libipt, ild: remove pti_set_map()
Almost all map assignments use a constant map argument.  The compiler may
optimize the error check away but this still leaves a set_error() in a void
function pti_set_map().

If we want to indicate errors in our return value, we have to either add a
return value to pti_set_map() to indicate the error or do the check at the
caller.

We do the latter and remove pti_set_map() in the process.

Change-Id: I9a7671a3807b0b422748dc7a89e40de84efd82aa
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>