libipt, ild: reduce operations in displacement decode
In disp_decode() and compute_disp_dec(), we check the map twice.  We also lookup
the disp table for opcode maps > 1, which we're going to ignore later on.

Move the checks from disp_decode() into compute_disp_dec() and rely on the
compiler to avoid the compute_disp_dec() result check in disp_decode() after
inlining the former into the latter.

Change-Id: I4505f8e132bf87f28c0114c2f658a63281fbef55
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>