libipt, image: bypass pt_msec_read
Instead of forwarding the read request from the image to the mapped section, we
directly read from the underlying section in pt_image_read().  This essentially
inlines pt_msec_read_mapped().

This removes pt_mapped_section as an additional layer that forwards requests to
the underlying section and reduces the amount of testing that would be required
if we added more request types.

Change-Id: I0b9f36db6b98c67d358c1a752d1123d93045aecb
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>