libipt, image: distinguish not-contained-in-section from other errors
In pt_image_read() we ignore sections we can't read from.  This does not
distinguish pte_nomap from other errors.

The former indicates that the section does not contain the address we try to
read from, which just means that we need to look for it in another section.
Other errors may indicate actual bugs.

Change it to fail in case of other errors.

Change-Id: I00b8826afb66023fd6936d93973c41261b9a6dfa
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>