libipt, image: refactor pt_image_read()
Narrow down the fast path to the least recently used section with caching
enabled.

Everything else, including disabled caching, goes onto the slow path.

Further simplify the slow path by doing the contained-in-section check in
pt_image.c and only map the section that is brought to the front of the LRU
list.  This requires a redundant check of the upper section bound.

Change-Id: Id2a0a6a370825f8270d814df7139881c09933e2a
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>