ptdump, ptxed: accept ':' in filenames
For most options that specify a filename, ptdump and ptxed accept an optional
suffix separated by ':' that specifies the file offset, a range in that file, or
the load address of an ELF file.

They simply search for the first occurrence of ':' and assume that everything
following it is such an optional suffix.  This is wrong for files that contain
':' as part of their name.

Change this to actually try to parse such an optional suffix starting at the
last occurrence of ':' in the filename argument.  If parsing fails, assume that
':' is part of the filename.

This fixes #7.

We still get it wrong for files that end with a valid suffix, e.g. "foo:42".
Without opening both "foo" and "foo:42" we won't know which specifies a valid
file.  And if both can be opened, we simply can't tell what the user meant.

Change-Id: I8e5c4741bcfd8bafadb9248a4ffa16b8a2ce6817
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>