libipt: fix end of trace indication
The upcoming end of the trace is indicated one instruction too late.  We don't
know whether the instruction retired.  Fix it.

Note that the instruction-flow decoder may still be used to decode past the end
of the trace until we reach an instruction for which trace is required in order
to proceed.

Change-Id: Ibd9af210c4ac3a8e9d278fe8b5cdb94639479a42
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>