libipt: process async branch event immediately
We indicate asynchronous branches in the preceding instruction by setting the
interrupted bit-field in struct pt_insn.  Processing of the async_branch event
is delayed until the next iteration.

This may cause us to miss a subsequent asynchronous disabled event on the
destination IP of the async branch.  Note that the instruction at that IP did
not retire.

Process the async_branch event immediately to catch this.

Change-Id: Ieedff32f040e75aec865de3b847846c4389185c9
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>