script: do not check for tools upfront in test.bash
The test.bash script checks whether the necessary tools are all available and
fails if one or more are not before executing any test.

This prevents the -g, -d, and -x options to be used for passing additional
arguments to the pttc, ptdump, and ptxed tool.

Remove the check.  This still causes tests to fail if one of the tools is not
availble but with a different error message.  If test.bash is called with more
than one test file, there will be one or more error messages per test.

Change-Id: I62f62cb546f1814ccba9078411107d39cd156a68
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>