libipt, time: don't store MTC offset
At TMA, we store the CTC bits below the MTC range in time->mtc_offset and the
CTC bits in the MTC range in time->ctc.

The latter is used to compute last_ctc at the first MTC after TMA.
The former is subtracted from the CTC delta between last and current CTC.

We might as well add the former to last_ctc before computing the CTC delta.
This allows us to merge time->mtc_offset into time->ctc.

Change-Id: Ib01b89a8ffa4796975fb7f26c8182a3d42096fcc
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>