ptxed: support truncated blocks
A block of instructions may be marked truncated if the last instruction in the
block can not be read entirely from the block's section.  It is instead provided
in the block itself.

Use the memory provided in the block in that case.

Our test system needs a few changes before we can write a test for this.

Change-Id: Ic9fcec2d32a908b13c8531ff2f6c0823241afefe
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>