ptxed: allow optional offset or range for raw binary files
Ptxed's --raw option takes a filename and a base address and loads the contents
of the former at the latter.

Extend this to further allow an optional file offset or file range to be
specified.  The base address is still mandatory.

This allows us to load parts of a test binary at different virtual addresses.
We may use this to create (identical) overlaps and thus split the test binary
into multiple image sections at arbitrary offsets.

Change-Id: I15093b66d9889e737d11ead58cf2a40579d67f4c
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>