libipt, block: do not mix cache fill and cache use
When filling the cache, we recursively call pt_blk_proceed_cached() to make use
of the cache after proceeding a single instruction.

Stop doing that.  Instead, check if there is a cache entry for the next IP and,
if not, recursively call pt_blk_proceed_fill_cache().

There shouldn't be any functional change but this new version will allow us to
bound the recursion more easily.

Change-Id: I6f479179c452d26e573d6922c8168e473b213759
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>