libipt, block: move block disabled indication out of pt_blk_apply_disabled()
When processing events, we set the disabled flag in pt_blk_apply_disabled() and
later on remove it again if the block had been empty.  Move setting the flag to
where we know we're going to need it.

As a side effect, this removes the block parameter from pt_blk_apply_disabled(),
which will be convenient in future patches.

Change-Id: I8a74afa2f2d268efb19bc0abdb9625d71469b5cd
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>