libipt, block: move interrupted indication out of pt_blk_apply_async_branch()
When processing events, we set the interrupted flag in
pt_blk_apply_async_branch() and later on remove it again if the block had been
empty.  Move setting the flag to where we know we're going to need it.

As a side effect, this removes the block parameter from
pt_blk_apply_async_branch(), which will be convenient in future patches.

Change-Id: I507649a2ac3cbe27a24722e0a01d5d114cad3e6a
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>