libipt, block: move abort/commit indication out of pt_blk_apply_tsx()
The tsx event handler pt_blk_apply_tsx() indicates an abort or commit in the
to-be-ended block unless traing is disabled or the block is empty.

In the proceed event flow, those conditions are already checked for other
reasons.  Move setting block flags to its callers to avoid duplicate checks.

As a side effect, this removes the block parameter from pt_blk_apply_tsx(), wich
will be convenient in future patches.

Change-Id: I56c5b1dd3c8a6c57b36b9e66878aa5e06ff180b8
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>