libipt, block: proceed to event with trace
The block decoder does not use trace when trying to reach an event location.
This works fine for most events but may cause it to fail with -pte_bad_query in
some cases.

For example, a MODE.EXEC binds to the next TIP and takes effect at the TIP IP.
In order to reach that location, we need to proceed past the (far) branch that
generated the TIP.

Change-Id: Ia56c37ea95c5d4e99f723bbfc79e5565c0e909e0
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>