libipt: indicate that tracing resumes in the enabled event
Add a new field 'resumed' to struct pt_event.variant.enabled to indicate whether
tracing resumes from the IP at which it had been disabled before.  This follows
the same semantics as the corresponding field in struct pt_insn and struct
pt_block.

This field can only be set by the instruction flow and block decoders since the
query decoder does not know the disable IP for synchronous disable events and is
further not able to account for calls.

The new field is currently always clear.  This patch prepares for instruction
flow and block decoder events by allowing tracing enabled events to be forwarded
to the user without losing the resumed indication.

Change-Id: I2f7998e79b152d8cb22ec38b94dc6267611af4c3
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>