libipt, insn: make insn argument in process_disabled_event() optional
Allow process_disabled_event() to be called with a NULL insn argument and skip
the insn indication in this case.

This will allow us to skip disabled events in an initial round of event
processing rather than skipping it in pt_insn_next() when processing the
subsequent enabled event after starting with a disabled event for the next
instruction.

Change-Id: I286ff208491b450d9c222793b0c2194f557ad501
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>