libipt, insn: add missing NULL pointer checks
Our policy is to check pointers in the function that dereferences them.  Such
checks are missing in process_async_disabled_event() and
process_sync_disabled_event().  Add them.

The respective pointers are checked in process_disabled_event() and
pt_insn_next_ip() so we're still OK - but we don't want to rely on that.

Change-Id: I8ea444858ff5dfe1e779a7bafdf7ae01c035244b
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>