libipt, insn: do not mix event processing before and after instruction decode
In process_events_after() we call process_events_before() again after each
processed event.  The idea was to process events that may appear before or after
and that may hide an after event.

Don't do that since process_events_before() may omit an indication in the
current instruction such as in a TIP.PGD, STOP sequence.  Instead, handle the
stop event in process_events_after() where we may indicate it in the
instruction.

We add a new test for the synchronous disable/stop case and rename existing
tests in that area to distinguish sync and async disables.

Change-Id: I2836f8b425e4aabdf53ad53beb609d62449b4e42
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>