libipt, insn: process peek events also when tracing is disabled
When tracing is disabled, we do not proceed to the next IP and hence do not
process peek events, either.

Process peek events nevertheless.  They will use a stale IP - which might
actually be the IP further down, e.g. in the next iteration of a loop.  That is,
it would be wrong if tracing were enabled, but since it isn't, we're fine.

We're not processing any event that would enable tracing again and even if we
did, such events would set the IP and we'd be fine, as well.

Change-Id: I779a286a010422c929dee33eeffb363e7cdca19d
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>