libipt, insn: do not ignore peek errors
We are ignoring errors from process_events_peek() arguing that we will run into
them again in the next iteration.

That is correct.  Yet, we already report proceed() errors one iteration early so
users are expected to look at @insn->iclass to determine whether the current
instruction is still valid and the error occured after it.

With that in place, there's not really a reason to delay reporting those errors.
This will just be in the way.  Report them immediately.

Change-Id: Ic090febe9e763f6f732317ff6def1e26bd12b6e5
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>