libipt, insn: allow process_events_after() to indicate events
Similar to process_events_peek(), we return a pts_status bit-vector from
process_events_peek().  This allows indicating user events and causes
process_events_peek() to be skipped in order to preserve the event that might
otherwise not be indicated again and would thus be lost or cause decode to fail.

Change-Id: I086e48c42e0a8e03a6730b8747a3acba65863146
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>