libipt, insn: process initial events
Instead of processing all initial events on the first call to pt_insn_next(),
already process them at the end of pt_insn_sync_<mode>().

This will allow us to indicate user events that precede the first instruction in
future patches.  As a side-effect, we now also get a pts_eos indication when
sychronizing on a PSB at the end of the trace.

Change-Id: I7702c3616c1a5a622eac1b6f5c3034f38e25bb29
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>