libipt, insn, block: forward enabled events to the user
Tracing enable is indicated by the respective flag in struct pt_insn on a call
to pt_insn_next().

This works fine as long as at least one instruction retires before the next
user-relevant event.  For events that occur after tracing has been enabled but
before the first instruction retires, it doesn't work.  We are not indicating
the event to the user.

For example, if tracing is enabled in the middle of a transaction (e.g. by
branching into a filter region) and the first instruction after enabling tracing
faults, control flows to the abort handler.  Both the transaction abort and the
control transfer are currently not indicated.  Tracing appears to start at the
abort handler.

Forward an enabled event to the user if there are more events pending.  We
really only need to do this if we would otherwise fail to indicate an event.
Making this distinction in the decoder is too expensive, though, so we rely on
the user to handle enabled events more often than strictly necessary.

The block decoder is able to return a block containing zero instructions but
indications for tracing enable, disable, resume, and others.  It would work
without the above changes.  We nevertheless do this also for the block decoder
to align the two decoders' behaviour.

This avoids having to special-case tests where one decoder might give a pair of
extra enable/disable events on the same location.  And it helps users who want
to allow using one or the other decoder or even switch between decoder types.

Change-Id: I97c65e8597b2d9bf0580ec6dc28425796dac581f
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>