libipt, block: process paging and vmcs as trailing event if tracing is disabled
Trailing event processing stops at synchronous paging and vmcs events.  They
will be processed in pt_blk_proceed_event() in the next iteration.  As long as
tracing is disabled, they will just be applied so we might as well process them
as trailing events in this case.

This shouldn't make a difference now but will become important if we want to
forward some events to the user.  They might hide behind a paging or vmcs event
and we may forget to indicate them to the user.

Change-Id: Iaf3aa1cc06b7fe0562ef48273cfd7e528b120db4
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>