libipt, insn: indicate enabled state
When returning a pts_status bit-vector, we indicate upcoming events as well as
the upcoming end of trace.  Also indicate whether tracing is currently enabled
or disabled.

This information is already available via the disabled field in struct pt_insn
provided by pt_insn_next().  It is missing in pt_insn_sync_*() and
pt_insn_event().

Change-Id: Ia790c19444e9a85b806d631295c77d7669c23003
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>