libipt, block: indicate enabled state
When returning a pts_status bit-vector, we indicate upcoming events as well as
the upcoming end of trace.  Also indicate whether tracing is currently enabled
or disabled.

This information is already available via the disabled field in struct pt_block
provided by pt_blk_next().  It is missing in pt_blk_sync_*() and pt_blk_event().

Change-Id: If49bea199f74a23de0b52e3f51b5782588b642c6
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>